refactor: extract anecdote list and voting into AnecdoteList component
Moved anecdote listing and voting logic from App.jsx to a dedicated AnecdoteList component. This improves modularity and separation of concerns, following the presentational/container pattern. App.jsx now acts as a clean container that composes AnecdoteForm and AnecdoteList, making the code more testable and maintainable.
